--- a/R/readAxivity.R
+++ b/R/readAxivity.R
@@ -366,7 +366,7 @@ readAxivity = function(filename, start = 0, end = 0, progressBar = FALSE, desire
     return(invisible(list(header = header, data = NULL)))
   }
   #############################################################################
-  # reinitiate file and start reading of data and sesrch the beginning of required
+  # reinitiate file and start reading of data and search the beginning of required
   seek(fid,0)
   # skip header
   seek(fid, 1024, origin = 'current')
@@ -403,9 +403,22 @@ readAxivity = function(filename, start = 0, end = 0, progressBar = FALSE, desire
     rawAccel = matrix(0, nrow = 300 * Npages, ncol = 6)
   }
   rawPos = 1
+  
+  # Pseudo code for possible revision:
+  # - Extract sector/block size in bytes => see c-link https://github.com/digitalinteraction/openmovement/blob/d8678127c8331196072215a2f1a3dfe7fa595915/Software/AX3/cwa-convert/c/main.c#L278
+  # - Op basis van block_start do seek() to skip
+  # - Run code as shown below
   for (i in 2:numDBlocks) {
-    raw = readDataBlock(fid, header_accrange = header$accrange, struc = struc,
-                        parameters = prevRaw$parameters)
+    timeSkip = start - prevRaw$start
+    blockDur = prevRaw$length / prevRaw$frequency
+    blockSkip = floor(timeSkip/blockDur) - 1
+    if (i >= blockSkip) { # start of recording
+      raw = readDataBlock(fid, header_accrange = header$accrange, struc = struc,
+                          parameters = prevRaw$parameters)
+    } else {
+      seek(fid, 512, origin = 'current') # skip block
+      next
+    }
     if (is.null(raw)) {
       break
     }
@@ -418,7 +431,6 @@ readAxivity = function(filename, start = 0, end = 0, progressBar = FALSE, desire
       prevRaw = raw
       next
     }
-    
     # Create array of times
     time = seq(prevStart, raw$start, length.out = prevLength + 1)
     
@@ -467,7 +479,6 @@ readAxivity = function(filename, start = 0, end = 0, progressBar = FALSE, desire
       break
     }
   }
-  
   #############################################################################
   # Process the last block of data if necessary
   if (pos <= nr) { # & ignorelastblock == FALSE) {