This PR speeds up function readAxivity for reading .cwa files (discussed in #14). 
The new speed-up is primarily gained by how the code searches for data chunks (batches) later on in the recording.

## Processing time improvements for running GGIR 2.9.1 - part 1 on:

### My own AX3 test file (284 MB):
- OLD: 400 seconds
- NEW: 102 seconds
- Difference: -75%

### My own AX6 test file (164 MB):
- OLD: 229 seconds
- NEW: 55 seconds
- Difference: -76%

### UK Biobank AX3 test file (265 MB):
- OLD: 350 seconds
- NEW: 92 seconds
- Difference: -74%

@jhmigueles Would you mind testing this on some of your own data to confirm these findings?

Thanks @egpbos for the brainstorm session this morning, as you can see it helped a lot.