@vincentvanhees 2023-05-12
> I'd be very interested to see also how the speed-up scales with the chunk size (my guess is that with smaller chunks the speedup is even larger) and with the data file size (again, larger files should mean more chunks, so more speed up).

Data is read block by block, with 1.2 seconds of sensor data per block, so I think this is already the minimum data length.